% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_mle_nr.R
\name{gamma_mle_nr}
\alias{gamma_mle_nr}
\title{Maximum Likelihood Estimation for Gamma Distribution Parameters}
\usage{
gamma_mle_nr(x, max_iter = 1000, tol = 1e-10)
}
\arguments{
\item{x}{A numeric vector of data points.}

\item{max_iter}{The maximum number of iterations for the Newton-Raphson algorithm. Default is 1000.}

\item{tol}{The convergence tolerance for the parameter estimates. Default is 1e-10.}
}
\value{
A list containing:
\describe{
  \item{alpha}{The estimated shape parameter of the Gamma distribution.}
  \item{beta}{The estimated scale parameter of the Gamma distribution.}
  \item{lhood}{The value of the log-likelihood function at the estimated parameters.}
}
}
\description{
Implements the Newton-Raphson algorithm for maximum likelihood estimation (MLE)
of the parameters of a Gamma distribution.
}
\examples{
set.seed(1234)
x <- rgamma(100, shape = 5, scale = 2)
res <- gamma_mle_nr(x, max_iter = 1000, tol = 1e-10)
print(res)
}
