% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gamma_fit.R
\name{plot_gamma_fit}
\alias{plot_gamma_fit}
\title{Plot Gamma Distribution Fit}
\usage{
plot_gamma_fit(x, alpha_est, beta_est)
}
\arguments{
\item{x}{A numeric vector representing the data.}

\item{alpha_est}{The estimated shape parameter of the Gamma distribution.}

\item{beta_est}{The estimated scale parameter of the Gamma distribution.}
}
\value{
A ggplot object showing the histogram and the fitted Gamma distribution.
}
\description{
This function generates a histogram of the data and overlays the fitted Gamma distribution density curve.
}
\examples{
set.seed(1234)
x <- rgamma(100, shape = 5, scale = 2)
alpha_est <- 5
beta_est <- 2
plot_gamma_fit(x, alpha_est, beta_est)
}
